refactor: carve out stripe-webhooks class
